+++ core/src/main/kotlin/com/github/spigotbasics/core/command/BasicsCommandExecutor.kt
@@ -1,5 +1,7 @@
 package com.github.spigotbasics.core.command
 
+import com.github.spigotbasics.core.command.raw.RawTabCompleter
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.messages.CoreMessages
 import com.github.spigotbasics.core.messages.MessageFactory
 import com.github.spigotbasics.core.module.BasicsModule
@@ -7,7 +9,7 @@ import com.github.spigotbasics.core.module.BasicsModule
 abstract class BasicsCommandExecutor(
     val coreMessages: CoreMessages,
     val messageFactory: MessageFactory,
-) : BasicsCommandContextHandler(), BasicsTabCompleter {
+) : BasicsCommandContextHandler(), RawTabCompleter {
     constructor(module: BasicsModule) : this(module.coreMessages, module.messageFactory)
 
     abstract fun execute(context: RawCommandContext): CommandResult?
