--- core/src/main/kotlin/com/github/spigotbasics/core/command/BasicsTabCompleter.kt
@@ -1,5 +1,0 @@
-package com.github.spigotbasics.core.command
-
-interface BasicsTabCompleter {
-    fun tabComplete(context: RawCommandContext): MutableList<String>?
-}
