+++ core/src/main/kotlin/com/github/spigotbasics/core/command/factory/CommandFactory.kt
@@ -1,7 +1,6 @@
 package com.github.spigotbasics.core.command.factory
 
 import com.github.spigotbasics.core.command.BasicsCommandManager
-import com.github.spigotbasics.core.command.RawCommandBuilder
 import com.github.spigotbasics.core.messages.CoreMessages
 import com.github.spigotbasics.core.messages.MessageFactory
 import org.bukkit.permissions.Permission
