+++ core/src/main/kotlin/com/github/spigotbasics/core/command/raw/RawTabCompleter.kt
@@ -1,0 +1,5 @@
+package com.github.spigotbasics.core.command.raw
+
+interface RawTabCompleter {
+    fun tabComplete(context: RawCommandContext): MutableList<String>?
+}
