+++ modules/basics-chat-format/src/main/kotlin/com/github/spigotbasics/modules/basicschatformat/commmands/ColorChatCommand.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicschatformat.commmands
 
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.extensions.addAnd
 import com.github.spigotbasics.core.extensions.partialMatches
 import com.github.spigotbasics.modules.basicschatformat.BasicsChatFormatModule
