+++ modules/basics-enderchest/src/main/kotlin/com/github/spigotbasics/modules/basicsenderchest/EnderchestCommand.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicsenderchest
 
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 
 class EnderchestCommand(private val module: BasicsEnderchestModule) : BasicsCommandExecutor(module) {
     override fun execute(context: RawCommandContext): CommandResult? {
