+++ modules/basics-gamemode/src/main/kotlin/com/github/spigotbasics/modules/basicsgamemode/GamemodeExecutor.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicsgamemode
 
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
 import org.bukkit.util.StringUtil
