+++ modules/basics-homes/src/main/kotlin/com/github/spigotbasics/modules/basicshomes/BasicsHomesModule.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicshomes
 
 import com.github.spigotbasics.common.Either
 import com.github.spigotbasics.core.Serialization
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.config.ConfigName
 import com.github.spigotbasics.core.module.AbstractBasicsModule
 import com.github.spigotbasics.core.module.loader.ModuleInstantiationContext
