+++ modules/basics-homes/src/main/kotlin/com/github/spigotbasics/modules/basicshomes/commands/DelHomeCommand.kt
@@ -3,7 +3,7 @@ package com.github.spigotbasics.modules.basicshomes.commands
 import com.github.spigotbasics.common.Either
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.extensions.partialMatches
 import com.github.spigotbasics.modules.basicshomes.BasicsHomesModule
 import org.bukkit.entity.Player
