+++ modules/basics-homes/src/main/kotlin/com/github/spigotbasics/modules/basicshomes/commands/SetHomeCommand.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicshomes.commands
 
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.extensions.getPermissionNumberValue
 import com.github.spigotbasics.core.extensions.partialMatches
 import com.github.spigotbasics.modules.basicshomes.BasicsHomesModule
