+++ modules/basics-weather/src/main/kotlin/com/github/spigotbasics/modules/basicsweather/WeatherCommand.kt
@@ -2,7 +2,7 @@ package com.github.spigotbasics.modules.basicsweather
 
 import com.github.spigotbasics.core.command.BasicsCommandExecutor
 import com.github.spigotbasics.core.command.CommandResult
-import com.github.spigotbasics.core.command.RawCommandContext
+import com.github.spigotbasics.core.command.raw.RawCommandContext
 import com.github.spigotbasics.core.extensions.partialMatches
 import org.bukkit.WeatherType
